refactor(VolumeRendering): thumbnailing logic
