+++ src/components/VolumeRendering.vue
@@ -2,11 +2,8 @@
 import {
   computed,
   defineComponent,
-  del,
   onBeforeUnmount,
-  reactive,
   ref,
-  set,
   watch,
   watchEffect,
 } from '@vue/composition-api';
@@ -15,21 +12,14 @@ import vtkPiecewiseWidget from '@/src/vtk/PiecewiseWidget';
 import { vtkSubscription } from '@kitware/vtk.js/interfaces';
 import vtkColorMaps from '@kitware/vtk.js/Rendering/Core/ColorTransferFunction/ColorMaps';
 import vtkPiecewiseFunctionProxy from '@kitware/vtk.js/Proxy/Core/PiecewiseFunctionProxy';
-import vtkLookupTableProxy from '@kitware/vtk.js/Proxy/Core/LookupTableProxy';
 import vtkColorTransferFunction from '@kitware/vtk.js/Rendering/Core/ColorTransferFunction';
 import ItemGroup from '@/src/components/ItemGroup.vue';
 import GroupableItem from '@/src/components/GroupableItem.vue';
-import vtkImageData from '@kitware/vtk.js/Common/DataModel/ImageData';
-import { Vector3 } from '@kitware/vtk.js/types';
 import { useResizeObserver } from '../composables/useResizeObserver';
 import { manageVTKSubscription } from '../composables/manageVTKSubscription';
-import { createVolumeThumbnailer } from '../core/thumbnailers/volume-thumbnailer';
 import { useCurrentImage } from '../composables/useCurrentImage';
-import { useCameraOrientation } from '../composables/useCameraOrientation';
-import { useImageStore } from '../store/datasets-images';
 import ColorFunctionSlider from './ColorFunctionSlider.vue';
 import { useVTKCallback } from '../composables/useVTKCallback';
-import { InitViewIDs, InitViewSpecs } from '../config';
 import { useViewConfigStore } from '../store/view-configs';
 import {
   getColorFunctionRangeFromPreset,
@@ -38,161 +28,12 @@ import {
   getShiftedOpacityFromPreset,
 } from '../utils/vtk-helpers';
 import { ColorTransferFunction } from '../types/views';
+import { useVolumeThumbnailing } from '../composables/useVolumeThumbnailing';
 
 const WIDGET_WIDTH = 250;
 const WIDGET_HEIGHT = 150;
 const THUMBNAIL_SIZE = 80;
 const TARGET_VIEW_ID = '3D';
-
-function resetOpacityFunction(
-  pwfProxy: vtkPiecewiseFunctionProxy,
-  dataRange: [number, number],
-  presetName: string
-) {
-  // reset pwf proxy range
-  pwfProxy.setDataRange(...dataRange);
-
-  const preset = vtkColorMaps.getPresetByName(presetName);
-  if (preset.OpacityPoints) {
-    const OpacityPoints = preset.OpacityPoints as number[];
-    const points = [];
-    for (let i = 0; i < OpacityPoints.length; i += 2) {
-      points.push([OpacityPoints[i], OpacityPoints[i + 1]]);
-    }
-
-    const [xmin, xmax] = dataRange;
-    const width = xmax - xmin;
-    const pointsNormalized = points.map(([x, y]) => [(x - xmin) / width, y]);
-
-    pwfProxy.setMode(vtkPiecewiseFunctionProxy.Mode.Points);
-    pwfProxy.setPoints(pointsNormalized);
-  } else {
-    pwfProxy.setMode(vtkPiecewiseFunctionProxy.Mode.Gaussians);
-    pwfProxy.setGaussians(vtkPiecewiseFunctionProxy.Defaults.Gaussians);
-  }
-}
-
-function useThumbnailing() {
-  const thumbnails = reactive<Record<string, Record<string, string>>>({});
-  const thumbnailer = createVolumeThumbnailer(THUMBNAIL_SIZE);
-  const currentThumbnails = ref<Record<string, string>>({});
-
-  const { currentImageMetadata, currentImageID, currentImageData } =
-    useCurrentImage();
-
-  // same as 3D view
-  const { cameraDirVec, cameraUpVec } = useCameraOrientation(
-    InitViewSpecs[InitViewIDs.Three].props.viewDirection,
-    InitViewSpecs[InitViewIDs.Three].props.viewUp,
-    currentImageMetadata
-  );
-
-  // used to interrupt a thumbnailing cycle if
-  // doThumbnailing is called again
-  let interruptSentinel = Symbol('interrupt');
-
-  async function doThumbnailing(imageID: string, image: vtkImageData) {
-    const localSentinel = Symbol('interrupt');
-    interruptSentinel = localSentinel;
-
-    thumbnailer.setInputImage(image);
-    const imageDataRange = image.getPointData().getScalars().getRange();
-
-    async function helper(presetName: string) {
-      // bail if a new thumbnail process has started
-      if (interruptSentinel !== localSentinel) {
-        return;
-      }
-
-      // sanity check; did the current image change
-      if (imageID !== currentImageID.value) {
-        return;
-      }
-
-      if (!(imageID in thumbnails)) {
-        set(thumbnails, imageID, {});
-      }
-
-      if (presetName in thumbnails[imageID]) {
-        return;
-      }
-
-      const opRange = getOpacityRangeFromPreset(presetName);
-      resetOpacityFunction(
-        thumbnailer.opacityFuncProxy,
-        opRange || imageDataRange,
-        presetName
-      );
-
-      thumbnailer.colorTransferFuncProxy.setMode(
-        vtkLookupTableProxy.Mode.Preset
-      );
-      thumbnailer.colorTransferFuncProxy.setPresetName(presetName);
-      const ctRange = getColorFunctionRangeFromPreset(presetName);
-      thumbnailer.colorTransferFuncProxy.setDataRange(
-        ...(ctRange || imageDataRange)
-      );
-
-      thumbnailer.resetCameraWithOrientation(
-        cameraDirVec.value as Vector3,
-        cameraUpVec.value as Vector3
-      );
-
-      const renWin = thumbnailer.scene.getRenderWindow();
-      renWin.render();
-      const imageURL = await renWin.captureImages()[0];
-      if (imageURL) {
-        set(thumbnails[imageID], presetName, imageURL);
-      }
-    }
-
-    PresetNameList.reduce(
-      (promise, presetName) => promise.then(() => helper(presetName)),
-      Promise.resolve()
-    );
-  }
-
-  // workaround for computed not properly working on deeply reactive objects
-  // in vue 2.
-  watch(
-    thumbnails,
-    () => {
-      if (currentImageID.value) {
-        currentThumbnails.value = thumbnails[currentImageID.value];
-      }
-    },
-    { deep: true }
-  );
-
-  // force thumbnailing to stop
-  onBeforeUnmount(() => {
-    interruptSentinel = Symbol('unmount');
-  });
-
-  // trigger thumbnailing
-  watch(
-    currentImageID,
-    (imageID) => {
-      if (imageID) {
-        doThumbnailing(imageID, currentImageData.value!);
-      }
-    },
-    { immediate: true }
-  );
-
-  // delete thumbnails if an image is deleted
-  const imageStore = useImageStore();
-  imageStore.$onAction(({ name, args, after }) => {
-    if (name === 'deleteData') {
-      const [id] = args as [string];
-      if (id in thumbnails) {
-        after(() => del(thumbnails, id));
-      }
-    }
-  });
-
-  return { currentThumbnails };
-}
 
 export default defineComponent({
   name: 'VolumeRendering',
@@ -376,7 +217,7 @@ export default defineComponent({
 
     // -- thumbnailing -- //
 
-    const { currentThumbnails } = useThumbnailing();
+    const { currentThumbnails } = useVolumeThumbnailing(THUMBNAIL_SIZE);
 
     // --- selection and updates --- //
